api/vcs: cleanup unused libgit2 methods after moving to go-git
Created with Sturdy
